refactor: extract config to a new file
