+++ src/libs/algolia/fetchAlgolia/index.ts
@@ -1,7 +1,8 @@
-export { attributesToRetrieve, fetchOffer, fetchOfferHits, fetchMultipleOffers } from './fetchOffer'
+export { fetchOffer, fetchOfferHits, fetchMultipleOffers } from './fetchOffer'
 export {
   filterOfferHit,
   isVenueHitTypeguard,
   transformOfferHit,
   useTransformOfferHits,
 } from './transformOfferHit'
+export { offerAttributesToRetrieve } from './config'
